+++ src/main/resources/static/css/menu.css
@@ -1,0 +1,387 @@
+@font-face {
+    font-family: 'Pretendard-Regular';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
+    font-weight: 400;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Pretendard-Light';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Light.woff') format('woff');
+    font-weight: 300;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Pretendard-Black';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Black.woff') format('woff');
+    font-weight: 900;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Pretendard-ExtraBold';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-ExtraBold.woff') format('woff');
+    font-weight: 800;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Pretendard-Bold';
+    src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Bold.woff') format('woff');
+    font-weight: 700;
+    font-style: normal;
+}
+body {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+    margin: 0;
+    font-family: 'Pretendard-Regular', Arial, sans-serif;
+    background-color: #fff;
+}
+
+header {
+    width: 100%;
+    background-color: #fff;
+    display: flex;
+    justify-content: center;
+    padding: 30px 0;
+}
+
+header .contentmenu {
+    padding-left: 200px;
+    display: flex;
+    gap: 20px;
+}
+
+header .contentmenu a {
+    text-decoration: none;
+    color: gray;
+    margin-right: 50px;
+    margin-left: 50px;
+    font-weight: bold;
+}
+
+.menu-link {
+    color: gray;
+    transition: color 0.3s ease;
+}
+
+.menu-link.active {
+    color: black;
+    font-weight: bold;
+}
+
+.content {
+    flex: 1;
+    margin-left: 200px;
+    padding: 20px;
+    display: flex;
+    justify-content: center; /* 중앙 정렬 */
+    align-items: center; /* 중앙 정렬 */
+}
+
+footer.footer, aside.sidebar {
+    background-color: #fff;
+}
+
+footer.footer {
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+    max-width: 767px;
+    display: flex;
+    justify-content: space-around;
+    border-top: 1px solid #ccc;
+    padding: 18px 0;
+    justify-content: center;
+}
+
+footer.footer .menu-item {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0 20px;
+}
+
+footer.footer .menu-item img {
+    width: 24px;
+    height: 24px;
+}
+
+aside.sidebar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    flex-direction: column;
+    align-items: flex-start;
+    height: 100%;
+    width: 200px;
+    max-width: none;
+    border-right: 1px solid #ccc;
+    padding-top: 20px;
+    padding-left: 20px;
+    box-sizing: border-box;
+}
+
+aside.sidebar a {
+    text-decoration: none;
+    display: flex;
+    align-items: center;
+    width: 100%;
+    padding: 10px;
+    box-sizing: border-box;
+    color: gray;
+}
+
+.logo {
+    margin-bottom: 10px;
+    display: flex;
+    justify-content: center;
+    width: 100%;
+}
+
+.logo img {
+    width: 50px;
+}
+
+.menu-item {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    margin: 0;
+    padding: 10px 0;
+    width: 100%;
+    box-sizing: border-box;
+}
+
+.menu-item img {
+    width: 24px;
+    height: 24px;
+    margin-right: 10px;
+}
+
+.menu-item.profile {
+    position: relative;
+}
+
+.menu-item.profile img {
+    border-radius: 50%;
+    border: 2px solid transparent;
+    box-shadow: none;
+    transition: box-shadow 0.3s ease;
+}
+
+.menu-item.profile.active img {
+    box-shadow: 0 0 0 2px black;
+}
+
+.sidebar .menu-item:last-child {
+    margin-bottom: auto;
+}
+
+.menu-item.active div {
+    font-weight: bold;
+}
+
+.menu-item div {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: black;
+}
+
+.menu-link.active div {
+    color: black;
+}
+
+.post-container {
+    display: flex;
+    gap: 20px;
+    align-items: flex-start;
+    flex-direction: column; /* 모바일 뷰에서 flex 방향을 컬럼으로 설정 */
+}
+
+.post {
+    background-color: #f9f9f9;
+    border: 1px solid #ddd;
+    padding: 15px;
+    border-radius: 10px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    flex: 1;
+    max-width: 500px;
+    width: 100%;
+    margin-bottom: 0; /* 하단 여백 제거 */
+    box-sizing: border-box;
+}
+
+.post-header {
+    display: flex;
+    align-items: center;
+    width: 100%;
+}
+
+.profile-pic {
+    border-radius: 50%;
+    width: 40px;
+    height: 40px;
+    margin-right: 10px;
+}
+
+.post-info {
+    display: flex;
+    flex-direction: column;
+}
+
+.username {
+    font-weight: bold;
+}
+
+.post-date {
+    font-size: 0.85em;
+    color: gray;
+}
+
+.post-images {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    grid-template-rows: repeat(2, 1fr);
+    gap: 5px;
+    width: 100%; /* 박스 너비에 맞추기 */
+    margin: 10px 0;
+}
+
+.post-image {
+    width: 100%;
+    height: 100%; /* 박스 높이에 맞추기 */
+    object-fit: cover;
+    border-radius: 5px;
+}
+
+.post-description {
+    margin: 10px 0;
+}
+
+.post-actions {
+    display: flex;
+    justify-content: space-between;
+    font-size: 0.9em;
+    color: gray;
+    width: 100%;
+}
+
+.like-text {
+    display: inline-block;
+    vertical-align: middle;
+    margin-left: 5px;
+    position: relative;
+    top: -2px;
+}
+
+.like-icon {
+    vertical-align: middle;
+    width: 20px;
+    height: 20px;
+    margin-right: 5px;
+    cursor: pointer;
+}
+
+.product-info {
+    border: 1px solid #ddd;
+    border-radius: 10px;
+    padding: 15px;
+    max-width: 300px;
+    align-self: flex-start;
+    overflow-x: auto;
+    width: 120%;
+    box-sizing: border-box;
+}
+.product-info table {
+    width: 100%;
+    border-collapse: collapse;
+}
+
+.product-info th, .product-info td {
+    padding: 5px;
+    text-align: left;
+}
+
+.product-info th {
+    background-color: #f2f2f2;
+}
+
+@media (min-width: 768px) {
+    footer.footer {
+        display: none;
+    }
+
+    footer.fixed-footer {
+        display: block;
+        justify-content: center;
+        align-items: center;
+        padding: 20px;
+        font-family: 'Pretendard-Light';
+        font-size: 10px;
+        color: #ccc;
+        background-color: #fff;
+        width: calc(100% - 200px);
+        margin-left: 200px;
+        position: relative;
+        box-sizing: border-box;
+        text-align: center;
+    }
+
+    .sidebar {
+        display: flex;
+    }
+}
+
+@media (max-width: 767px) {
+    .content {
+        margin-left: 0;
+        flex-direction: column;
+        align-items: center;
+        overflow-y: auto;
+        height: calc(100vh - 100px); /* 푸터 높이만큼 빼줌 */
+        padding-bottom: 100px; /* 푸터 높이만큼 여유 공간 추가 */
+    }
+
+    .sidebar {
+        display: none;
+    }
+
+    footer.fixed-footer {
+        display: none;
+    }
+
+    header .contentmenu {
+        padding-left: 0;
+        display: flex;
+        justify-content: center;
+        gap: 10px;
+    }
+
+    header .contentmenu a {
+        margin-right: 10px;
+        margin-left: 10px;
+    }
+
+    .product-info {
+        max-width: 100%;
+        margin-top: 20px;
+    }
+
+    .post-container {
+        flex-direction: column;
+        align-items: center;
+    }
+
+    .product-info.mobile-view {
+        display: block;
+        max-width: 100%;
+        width: 100%;
+    }
+}
